refactor: Split and factorized code
