+++ MailCore/Models/Contact/CommonContact.swift
@@ -18,122 +18,9 @@
 
 import Foundation
 import InfomaniakCore
-import InfomaniakDI
 import MailResources
 import Nuke
 import UIKit
-
-public struct AvatarImageRequest {
-    let imageRequest: ImageRequest?
-    let shouldAuthenticate: Bool
-
-    public func authenticatedRequestIfNeeded(token: ApiToken) -> ImageRequest? {
-        guard let unauthenticatedImageRequest = imageRequest,
-              let unauthenticatedUrlRequest = unauthenticatedImageRequest.urlRequest else {
-            return nil
-        }
-
-        guard shouldAuthenticate else {
-            return unauthenticatedImageRequest
-        }
-
-        var authenticatedUrlRequest = unauthenticatedUrlRequest
-        authenticatedUrlRequest.addValue(
-            "Bearer \(token.accessToken)",
-            forHTTPHeaderField: "Authorization"
-        )
-
-        return ImageRequest(urlRequest: authenticatedUrlRequest)
-    }
-}
-
-public enum ContactConfiguration: CustomDebugStringConvertible {
-    public var debugDescription: String {
-        switch self {
-        case .recipient(let recipient, _):
-            return ".recipient:\(recipient.name) \(recipient.email)"
-        case .user(let user):
-            return ".user:\(user.displayName) \(user.email)"
-        case .contact(let contact):
-            return ".contact:\(contact.fullName) \(contact.email)"
-        case .emptyContact:
-            return ".emptyContact"
-        }
-    }
-
-    case recipient(recipient: Recipient, contextMailboxManager: MailboxManager)
-    case user(user: UserProfile)
-    case contact(contact: CommonContact)
-    case emptyContact
-}
-
-/// A standard cache for `CommonContact`, used by DI
-public typealias ContactCache = NSCache<NSNumber, CommonContact>
-
-/// Creating a `CommonContact` is expensive, relying on a cache to reduce hangs
-public enum CommonContactCache {
-    /// The underlying standard cache
-    @LazyInjectService private static var cache: ContactCache
-
-    /// Get a contact from cache if any or nil
-    public static func getContactFromCache(contactConfiguration: ContactConfiguration) -> CommonContact? {
-        let key: NSNumber
-        switch contactConfiguration {
-        case .recipient(let recipient, _):
-            key = NSNumber(value: recipient.id.hash)
-        case .user(let user):
-            key = NSNumber(value: user.id)
-        case .contact(let contact):
-            return contact
-        case .emptyContact:
-            return CommonContact.emptyContact
-        }
-
-        return cache.object(forKey: key)
-    }
-
-    /// Get a contact from cache or build it
-    public static func getOrCreateContact(contactConfiguration: ContactConfiguration) -> CommonContact {
-        let contact: CommonContact
-        let key: NSNumber
-
-        switch contactConfiguration {
-        case .recipient(let recipient, let contextMailboxManager):
-            key = NSNumber(value: recipient.id.hash)
-            contact = getOrCreateContact(recipient: recipient, contextMailboxManager: contextMailboxManager, key: key)
-        case .user(let user):
-            key = NSNumber(value: user.id)
-            contact = getOrCreateContact(user: user, key: key)
-        case .contact(let wrappedContact):
-            key = NSNumber(value: wrappedContact.id)
-            contact = wrappedContact
-        case .emptyContact:
-            contact = CommonContact.emptyContact
-            key = NSNumber(value: contact.id)
-        }
-
-        cache.setObject(contact, forKey: key)
-        return contact
-    }
-
-    private static func getOrCreateContact(user: UserProfile, key: NSNumber) -> CommonContact {
-        guard let contact = CommonContactCache.cache.object(forKey: key) else {
-            return CommonContact(user: user)
-        }
-
-        return contact
-    }
-
-    private static func getOrCreateContact(recipient: Recipient,
-                                           contextMailboxManager: MailboxManager,
-                                           key: NSNumber) -> CommonContact {
-        guard let contact = CommonContactCache.cache.object(forKey: key) else {
-            return CommonContact(recipient: recipient, contextMailboxManager: contextMailboxManager)
-        }
-
-        return contact
-    }
-}
 
 public final class CommonContact: Identifiable {
     /// Empty contact is a singleton
